+++ src/items/mod.rs
@@ -8,30 +8,26 @@ use crate::{
     interaction::InteractionController,
     margin::{Margin, MarginExt},
     selection_indicator::{style::IndicatorStyle, SelectionIndicatorController},
-    MenuItem, MenuStyle,
+    MenuStyle,
 };
 use embedded_graphics::{
     draw_target::DrawTarget,
     pixelcolor::Rgb888,
     prelude::{PixelColor, Point, Size},
-    primitives::{Rectangle, StyledDrawable},
+    primitives::Rectangle,
     text::{renderer::TextRenderer, Baseline},
     Drawable,
 };
 use embedded_layout::prelude::*;
 use embedded_text::{alignment::HorizontalAlignment, style::TextBoxStyleBuilder, TextBox};
 
-pub struct MenuLine<I> {
+pub struct MenuLine {
     bounds: Margin<Rectangle>,
     value_width: u32,
-    item: I,
 }
 
-impl<I> MenuLine<I>
-where
-    I: MenuItem,
-{
-    pub fn new<C, S, IT, P>(item: I, style: MenuStyle<C, S, IT, P>) -> MenuLine<I>
+impl MenuLine {
+    pub fn new<C, S, IT, P>(longest_value: &str, style: &MenuStyle<C, S, IT, P>) -> Self
     where
         C: PixelColor,
         S: IndicatorStyle,
@@ -40,8 +36,6 @@ where
     {
         let style = style.text_style();
 
-        let longest_value = item.longest_value_str();
-
         let value_width = style
             .measure_string(longest_value, Point::zero(), Baseline::Top)
             .bounding_box
@@ -49,7 +43,6 @@ where
             .width;
 
         MenuLine {
-            item,
             value_width,
             bounds: Rectangle::new(
                 Point::zero(),
@@ -59,43 +52,26 @@ where
         }
     }
 
-    pub fn as_item(&self) -> &I {
-        &self.item
+    pub fn empty() -> Self {
+        MenuLine {
+            bounds: Rectangle::new(Point::zero(), Size::new(1, 1)).with_margin(0, 0, -1, 0),
+            value_width: 0,
+        }
     }
 
-    pub fn as_item_mut(&mut self) -> &mut I {
-        &mut self.item
-    }
-}
-
-impl<I> View for MenuLine<I> {
-    fn translate_impl(&mut self, by: Point) {
-        self.bounds.translate_mut(by);
-    }
-
-    fn bounds(&self) -> Rectangle {
-        self.bounds.bounds()
-    }
-}
-
-impl<C, S, I, IT, P> StyledDrawable<MenuStyle<C, S, IT, P>> for MenuLine<I>
-where
-    C: PixelColor + From<Rgb888>,
-    I: MenuItem,
-    S: IndicatorStyle,
-    IT: InteractionController,
-    P: SelectionIndicatorController,
-{
-    type Color = C;
-    type Output = ();
-
-    fn draw_styled<D>(
+    pub fn draw_styled<D, C, S, IT, P>(
         &self,
+        title: &str,
+        value_text: &str,
         style: &MenuStyle<C, S, IT, P>,
         display: &mut D,
-    ) -> Result<Self::Output, D::Error>
+    ) -> Result<(), D::Error>
     where
-        D: DrawTarget<Color = Self::Color>,
+        D: DrawTarget<Color = C>,
+        C: PixelColor + From<Rgb888>,
+        S: IndicatorStyle,
+        IT: InteractionController,
+        P: SelectionIndicatorController,
     {
         let text_bounds = self.bounds.bounds();
         let display_area = display.bounding_box();
@@ -105,7 +81,6 @@ where
         }
 
         let text_style = style.text_style();
-        let value_text = self.item.value();
 
         let mut inner_bounds = self.bounds.inner.bounds();
         inner_bounds.size.width = display_area.size.width - self.bounds.left as u32;
@@ -120,8 +95,18 @@ where
         .draw(display)?;
 
         inner_bounds.size.width -= self.value_width;
-        TextBox::new(self.item.title(), inner_bounds, text_style).draw(display)?;
+        TextBox::new(title, inner_bounds, text_style).draw(display)?;
 
         Ok(())
+    }
+}
+
+impl View for MenuLine {
+    fn translate_impl(&mut self, by: Point) {
+        self.bounds.translate_mut(by);
+    }
+
+    fn bounds(&self) -> Rectangle {
+        self.bounds.bounds()
     }
 }