+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Unreleased (??)
 
  - Removed `SingleTouch::new`.
  - Fields of `SingleTouch` are now public.
+ - **breaking**: Removed `MenuLine` from the menu type signature.
+ - **breaking**: `MenuLine` is no longer generic.
+ - Added `MenuItem::set_style` which is used internally to set up a menu item.
 
 0.2.0 (2023-06-23)
 ==================